requirements.txt
python-dotenv==1.0.0

# requests
requests==2.31.0
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
urllib3==2.0.2
# pandas
pandas==2.0.2
numpy==1.24.3 
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
# flask
flask==2.3.2
PyNaCl==1.5.0
Werkzeug==2.3.3
click==8.1.3
# Flask-Migrate -> dealing with updating database tables in python then porting the changes to SQL
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Mako==1.2.4
SQLAlchemy==2.0.12
alembic==1.10.4
# Flask-WTF provides a convenient and secure way to create and manage web forms in Flask applications., WTForms is a popular Python library for handling web forms. It provides form validation, CSRF (Cross-Site Request Forgery) protection, and other useful features when working with forms in web applications.
WTForms==2.2.1
flask_wtf==1.1.1

# PostgreSQL database adapter for the Python programming language. 
psycopg2==2.9.6

# needed for heroku deployment
gunicorn==20.0.4